+++ apps/sync/config/config.exs
@@ -1,6 +1,7 @@
 import Config
 
 config :chalk_sync,
+  dev_tools: false,
   port: 4100,
   stateholder: ChalkSync.Stateholder.Memory,
   token_verifier: ChalkSync.Auth.DevTokenVerifier