+++ apps/sync/config/dev.exs
@@ -1,1 +1,3 @@
 import Config
+
+config :chalk_sync, dev_tools: true